Continuing with Modal OrderDetails. WIP
